pyproject.toml
[project]
name = "soundtools"
version = "0.3.0"
description = "Personalized wake word detection using Whisper Tiny encoder"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "sounddevice",
    "openai-whisper>=20250625",
    "torch>=2.11.0",
]

[tool.uv]
override-dependencies = [
    "tflite-runtime ; python_version > '99'",
]
